Add update-account governance job and add validation

+++ tests/bos.sh
@@ -18,6 +18,8 @@
 export script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 # For parallel when default shell is not bash (we need exported functions)
 export SHELL=$(which bash)
+export job_log="$script_dir/bos-test-log.txt"
+export test_output="$script_dir/bos-test-output.txt"
 
 source "$script_dir/test_runner.sh"
 
@@ -37,8 +39,9 @@ run_test(){
   cat readme.md
   echo
 
-
-  bos_cmd="${bos_bin} --chain-url='$BURROW_HOST:$BURROW_GRPC_PORT' --address '$key1_addr' --set 'addr1=$key1_addr' --set 'addr2=$key2_addr' --set 'addr2_pub=$key2_pub'"
+#
+  bos_cmd="${bos_bin} --chain-url=$BURROW_HOST:$BURROW_GRPC_PORT --keys=$BURROW_HOST:$BURROW_GRPC_PORT \
+   --address $key1_addr --set addr1=$key1_addr --mempool-signing --set addr2=$key2_addr --set addr2_pub=$key2_pub"
   [[ "$debug" == true ]] && bos_cmd="$bos_cmd --debug"
   echo "executing bos with command line:"
   echo "$bos_cmd"
@@ -84,10 +87,14 @@ bos_tests(){
   # Cleanup
   cleanup() {
     goto_base
-    git clean -fdxq
-    if [[ "$test_exit" -eq 0 ]]
+
+    if [[ "$clean" == true ]]
     then
-        rm -f "$job_log" "$test_output"
+      git clean -fdxq
+      if [[ "$test_exit" -eq 0 ]]
+      then
+          rm -f "$job_log" "$test_output"
+      fi
     fi
     # This exits so must be last thing called
     test_teardown
